chore: update bench
