+++ lact-gui/Cargo.toml
@@ -20,8 +20,8 @@ tracing = { workspace = true }
 anyhow = { workspace = true }
 tracing-subscriber = { workspace = true }
 chrono = { workspace = true }
+gtk = { workspace = true, features = ["v4_6", "blueprint"] }
 
-gtk = { version = "0.9", package = "gtk4", features = ["v4_6", "blueprint"] }
 adw = { package = "libadwaita", version = "0.7.1", features = [
     "v1_4",
 ], optional = true }
